Dont return an HTTP error code reflecting the JMX operation's result,
but only when something fails on the uppermost agent level. This implies, that an HTTP
code of 200 can be returned even if one or more JMX operations failed. So it is necessary
to check both, the HTTP status code for errors on the transport and uppermost agent level
and the status code contained in the Jolokia responsed (key "status" in the JSON response
object).