‟Stashed” the http server url on the system properties for agents
Dynamic attaching agents require the ability to determine if the agent has
already been attached, and if so see if they need to attach a new agent
(currently an error) or use a pre-existing, pre-loaded agent.

This is only done for agents that are attached via the JVM agent interfaces,
if an application has jolokia already bound through other means, then this
will not be reflected in the system properties.

For applications that use Dynamic Attach to run jolokia, this means that the
following code can be used to see if a http server is already running, and as
such use that

   --- %< ---
   String vmHandle = "12345";
   VirtualMachine vm = VirtualMachine.attach(vmHandle);

   try {
     String agentUrl = (String) vm.getSystemProperties().get("jolokia.agent_url");
     if (agentUrl == null) {

         URI jarUri = JvmAgentJdk6.class
           .getProtectionDomain()
           .getCodeSource()
           .getLocation()
           .toURI();

         String agent = new File(jarUri).getAbsolutePath();
         vm.loadAgent(AGENT, "port=7788,host=localhost,agentContext=/jmx");
         System.out.println("Attached JMX Rest bridge to " + vmHandle);
     } else {
         System.out.println("Attach already exists " + agentUrl);
     }
   } finally {
     vm.detach();
   }
   --- >% ---

Also made dynamic attach throw exceptions rather than raise stdout, as this is
easier for those clients to work with.