Added lazy initialization for the BackendManager which can now be done on the first request.
If the agent is installed via -javaagent: startup option then the BackendManager is initialized lazily,
This is necessary in order to given the JVM time to start everything else, i.e. those things which are required
for the detectors to detect the proper server versions. The implementation should not influence much the regular
behaviours (except for an added reference check for null, which most of the time is true and dont add any
extra work)