Worked on the pull notification backend.
Also, introduced a Jolokia id for uniquely identifying a agent within a JVM.
This can be used to create agent specific MBeans. However, the id will
change over restarts, so it is not so suitable for static configuration
scenarios. The notification pull backend e.g. will use this ID for creating
the MBean holding the notification store. The name of this Mbean is returned
during the register call. The ID is also available now from the version command.