Fixes #108 by providing an alternative way for adding an Authenticator
A custom authenticator can be added by subclassing JolokiaServerConfig and
overriding getAuthenticator(). By default, if a user and a password is given, a
BasiAuthenticator checking this credentials is used. I decided to do it that way,
to keep the JolokiaServer immutable with respect to its initialization (a set method
could be invoked later, too) and to keep it 'sealed' after construction.

Also, JolokiaServer has now a new constructor for using an existing HttpServer instead
of creating one on its own. If a HttpServer is given that way, it won't be started or
stopped with JolokiaServer's lifecycle method.