Migrated ServerDetector to be JolokiaServices, too.
Note: The SortedSet based storage of services should probably moved
to a List based one if we want to allow services with the same order (for
a SortedSet the same order implies "equality", which allows only one
service per order be stored. The LocalServiceFactory guarantees a unique
order, so thats not a problem. But probably when getting services from another
subsystem like OSGi).