Fix race condition wih date headers
Currently the date format for creating header dates where cached in a member variable. However, since SimpleDateFormat is not thread safe it is now recreated for every request (wonder if it would be more effective to reuse the date format and put it into a synchronized method. Probably not). Fixes #184