Fix setting of "secure" flag for autodiscovery.
Now this is set after the first request has been served by the agent. The reason is, that the security information is only obtainable from a HttpRequest object. So there might be small window, where this information is not set, in which case this property is missing the auto discovery response. Fixes #193.