No default truncation of collections for WAR agent
Truncation of collections can be tuned with the 'maxCollectionSize' config parameter. The default was infinite for the JVM agent and Mule agent but 1000 for the WAR agent. This turned out to be too less (see https://github.com/hawtio/hawtio/issues/1725#issuecomment-81602811). Now the limitation has been removed, which shouldn't be a big problem for most. You can always add your own limitation by setting the parameter in web.xml. Fixes #187.