Remove transitive dependencies on httpclient-osgi
The Jolokia Java client uses httpclient-osgi, which is an 'fat' OSGi bundle with all dependencies inlined. However, these deps are still declared in the pom, so classes ende up twice in the classpath for e.g. IDEs which build up the classpath from the Maven pom. The dependencies are now exclude, however there might be transitive dependencies in deeper levels which must be included explicitely now. Like the commons-logging which are needed for the tests.

Fixes #180