#234 : Add new option 'allowDnsReverseLookup'
During a restrictor check by default the hostname is determined which might cause a reverse DNS lookup. In order to avoid this (e.g. because there is nor restriction on the hostname), this configuration property can be set to `false`.
For backwards compatibility it is set to `true` but will change to `false` in Jolokia 2.0