correct the sample application
The sample jolokia application in the tutorial demo seems to have
drifted from the original libraries that it relies on. The
`J4pReadResponse` returned from the client execution does not return
`Map<String,String>` values as expected, but instead returns a
`Map<String,Long>`. This change impacted the string conversion lines
that follow as well. The updated code was tested with the linked jar
files and OpenJDK 1.8.0_111.