feature(osgi-agent): Update authMode to support "service-any" and "service-all"
service-any: A single Authenticator is enough to succeed
service-all: All Authenticator services need to be positive

Also, switched "Authenticator" from an abstract base class to an
interface. So existing code might need a recompile, not sure.
But since the security package was not exposed previously, I don't think
this is an issue.