Have KeyStoreUtil.updateWithServerPems load the full cert chain
updateWithServerPems uses CertificateFactory.GenerateCertificate(), which only loads the
first certificate in the server certificate file. When using a certificate that is signed by
an intermediate CA (like Let's Encrypt,) this would result in a certifiacte that won't be
accepted by clients that only has the root CA, and not the intermediate CA, in their trusted
certificates stores.

This replaces that call with a call to CertificateFactory.GenerateCertificates(), that
loads the full chain, allowing clients with just the root chain to trust the server.

Signed-off-by: Shay Elkin <shay@elkin.io>